Allow the ContentFile mimetype to differ from the file's mimetype
The mechanism for uploading a file and getting it converted
automatically involves the equivalent of:

    CreateFile({ 'mimetype': some_Google_docs_mime_type })
    SetContentFile(file_with_different_mime_type)

Assuming the uploaded blob has a compatible MIME type, Drive attempts to
convert it to the target GSuite format.

This patch allows the GoogleDriveFile object to have a different
mimetype than the file uploaded via SetContentFile in order to trigger
this conversion.